+++ apps/api/src/cli.ts
@@ -2,8 +2,6 @@ import { BootstrapConsole } from 'nestjs-console';
 
 import { AppModule } from './app/app.module';
 
-import { Logger } from '@tractr/nestjs-core';
-
 const bootstrap = new BootstrapConsole({
   module: AppModule,
   useDecorators: true,
@@ -14,11 +12,9 @@ const bootstrap = new BootstrapConsole({
 
 /* eslint-disable @typescript-eslint/no-floating-promises */
 bootstrap.init().then(async (app) => {
-  const logger = new Logger();
   let exitCode = 0;
   try {
     await app.init();
-    app.useLogger(logger);
     await bootstrap.boot();
   } catch (e) {
     console.error(e);