+++ libs/nestjs/request-timestamp/src/lib/index.ts
@@ -1,0 +1,2 @@
+export * from './request-timestamp.module';
+export * from './request-timestamp.middleware';