+++ libs/terraform/component/deployment-trigger/src/lib/index.ts
@@ -1,0 +1,4 @@
+export * from './interfaces';
+export * from './deployment-trigger.component';
+export * from './deployment-trigger-role.component';
+export * from './deployment-trigger-event.component';