+++ libs/terraform/group/network/src/lib/network.group.ts
@@ -1,9 +1,13 @@
-import { vpc } from '@cdktf/provider-aws';
-
 import { BaseComponent, PrivateComponent, PublicComponent } from './components';
 import { NETWORK_GROUP_DEFAULT_CONFIG } from './configs';
 import {
-  AvailabilityZoneInfo,
+  getCidrBlockForIndex,
+  getPrivateSubnetNumber,
+  getPublicSubnetNumber,
+  getZoneName,
+} from './helpers';
+import {
+  NetworkGroupArtifacts,
   NetworkGroupConfig,
   NetworkGroupDefaultConfig,
 } from './interfaces';
@@ -11,31 +15,45 @@ import {
 import {
   AwsComponent,
   AwsProviderConstruct,
+  mergeConfigurations,
 } from '@tractr/terraform-component-aws';
 
 export class NetworkGroup extends AwsComponent<
-  NetworkGroupConfig & NetworkGroupDefaultConfig
+  NetworkGroupConfig & NetworkGroupDefaultConfig,
+  NetworkGroupArtifacts
 > {
-  protected readonly baseComponent: BaseComponent;
-
-  protected readonly publicComponents: PublicComponent[];
-
-  protected readonly privateComponents: PrivateComponent[];
-
-  protected readonly availabilityZonesInfo: AvailabilityZoneInfo[] = [];
-
   constructor(
     scope: AwsProviderConstruct,
     id: string,
     config: NetworkGroupConfig,
   ) {
-    super(scope, id, { ...NETWORK_GROUP_DEFAULT_CONFIG, ...config });
-    this.baseComponent = this.createBaseComponent();
-    // Create one public and one private subnet in each zone
-    this.publicComponents = this.createPublicComponents();
-    this.privateComponents = this.createPrivateComponents();
+    super(scope, id, mergeConfigurations(NETWORK_GROUP_DEFAULT_CONFIG, config));
   }
 
+  protected createComponents(): void {
+    // Create network base (VPC, Internet Gateway, Route Table, etc)
+    const baseComponent = this.createBaseComponent();
+    // Create one public and one private subnet in each zone
+    const publicComponents = this.createPublicComponents(baseComponent);
+    const privateComponents = this.createPrivateComponents(
+      baseComponent,
+      publicComponents,
+    );
+    // Populate artifacts
+    this.artifacts = {
+      ...baseComponent.artifacts,
+      publicSubnets: publicComponents.map((c) => c.artifacts.subnet),
+      privateSubnets: privateComponents.map((c) => c.artifacts.subnet),
+      availabilityZonesInfo: this.config.zones.map((letter) => ({
+        fullName: getZoneName(this.provider, letter),
+        shortName: letter,
+      })),
+    };
+  }
+
+  /**
+   * Create the base component (VPC, Internet Gateway, Route Table, etc)
+   */
   protected createBaseComponent() {
     return new BaseComponent(this, 'base', {
       cidrPrefix: this.config.cidrPrefix || '172.16',
@@ -43,72 +61,59 @@ export class NetworkGroup extends AwsComponent<
     });
   }
 
-  protected createPublicComponent(index: number) {
-    const zoneLetter = this.config.zones[index];
-    const publicId = `pub-${zoneLetter}`;
-    return new PublicComponent(this, publicId, {
-      availabilityZone: this.baseComponent.getZoneName(zoneLetter),
-      vpcId: this.getVpcIdAsToken(),
-      cidrBlock: this.baseComponent.getPublicCidrBlockForIndex(index),
-      ipv6CidrBlock: this.baseComponent.getPublicIpv6CidrBlockForIndex(index),
+  /**
+   * Create public subnets
+   */
+  protected createPublicComponents(baseComponent: BaseComponent) {
+    return this.config.zones.map((zone, index) => {
+      const zoneLetter = this.config.zones[index];
+      const publicId = `pub-${zoneLetter}`;
+      const subnetNumber = getPublicSubnetNumber(index);
+
+      return new PublicComponent(this, publicId, {
+        availabilityZone: getZoneName(this.provider, zoneLetter),
+        vpc: baseComponent.artifacts.vpc,
+        cidrBlock: getCidrBlockForIndex(
+          baseComponent.artifacts.vpcData.cidrBlock,
+          subnetNumber,
+        ),
+        ipv6CidrBlock: getCidrBlockForIndex(
+          baseComponent.artifacts.vpcData.ipv6CidrBlock,
+          subnetNumber,
+        ),
+      });
     });
   }
 
-  protected createPublicComponents() {
-    return this.config.zones.map((zone, index) =>
-      this.createPublicComponent(index),
-    );
-  }
-
-  protected createPrivateComponent(
-    index: number,
-    publicComponent: PublicComponent,
+  /**
+   * Create private subnets
+   */
+  protected createPrivateComponents(
+    baseComponent: BaseComponent,
+    publicComponents: PublicComponent[],
   ) {
-    const zoneLetter = this.config.zones[index];
-    const privateId = `prv-${zoneLetter}`;
-    return new PrivateComponent(this, privateId, {
-      publicSubnetId: publicComponent.getSubnetIdAsToken(),
-      egressOnlyInternetGatewayId:
-        this.baseComponent.getEgressOnlyInternetGatewayGatewayIdAsToken(),
-      internetGatewayId: this.baseComponent.getInternetGatewayIdAsToken(),
-      internetAccessMode: this.config.internetAccessMode,
-      availabilityZone: this.baseComponent.getZoneName(zoneLetter),
-      vpcId: this.getVpcIdAsToken(),
-      cidrBlock: this.baseComponent.getPrivateCidrBlockForIndex(index),
-      ipv6CidrBlock: this.baseComponent.getPrivateIpv6CidrBlockForIndex(index),
+    return publicComponents.map((publicComponent, index) => {
+      const zoneLetter = this.config.zones[index];
+      const privateId = `prv-${zoneLetter}`;
+      const subnetNumber = getPrivateSubnetNumber(index);
+
+      return new PrivateComponent(this, privateId, {
+        publicSubnet: publicComponent.artifacts.subnet,
+        egressOnlyInternetGateway:
+          baseComponent.artifacts.egressOnlyInternetGateway,
+        internetGateway: baseComponent.artifacts.internetGateway,
+        internetAccessMode: this.config.internetAccessMode,
+        availabilityZone: getZoneName(this.provider, zoneLetter),
+        vpc: baseComponent.artifacts.vpc,
+        cidrBlock: getCidrBlockForIndex(
+          baseComponent.artifacts.vpcData.cidrBlock,
+          subnetNumber,
+        ),
+        ipv6CidrBlock: getCidrBlockForIndex(
+          baseComponent.artifacts.vpcData.ipv6CidrBlock,
+          subnetNumber,
+        ),
+      });
     });
-  }
-
-  protected createPrivateComponents() {
-    return this.publicComponents.map((publicComponent, index) =>
-      this.createPrivateComponent(index, publicComponent),
-    );
-  }
-
-  getPublicSubnetsIdsAsTokens(): string[] {
-    return this.publicComponents.map((component) =>
-      component.getSubnetIdAsToken(),
-    );
-  }
-
-  getPrivateSubnetsIdsAsTokens(): string[] {
-    return this.privateComponents.map((component) =>
-      component.getSubnetIdAsToken(),
-    );
-  }
-
-  getVpcIdAsToken(): string {
-    return this.baseComponent.getVpcIdAsToken();
-  }
-
-  getInternetGateway(): vpc.InternetGateway {
-    return this.baseComponent.getInternetGateway();
-  }
-
-  getAvailabilityZonesInfo(): AvailabilityZoneInfo[] {
-    return this.config.zones.map((letter) => ({
-      fullName: this.baseComponent.getZoneName(letter),
-      shortName: letter,
-    }));
   }
 }