+++ libs/terraform/group/pool/src/lib/interfaces/ecs.interface.ts
@@ -1,15 +1,28 @@
+import {
+  cloudwatch,
+  ecs,
+  elb,
+  s3,
+  secretsmanager,
+  vpc,
+} from '@cdktf/provider-aws';
+
+import { PrivateDnsComponent } from '@tractr/terraform-component-private-dns';
 import { DockerApplications } from '@tractr/terraform-group-registry';
-import { ReverseProxyComponentPublicConfig } from '@tractr/terraform-service-reverse-proxy';
+import { ExecutionRoleComponent } from '@tractr/terraform-service-ecs';
+import {
+  ReverseProxyComponent,
+  ReverseProxyComponentPublicConfig,
+  ReverseProxyTaskRoleComponent,
+} from '@tractr/terraform-service-reverse-proxy';
 
 export interface EcsComponentInternalConfig {
-  subnetsIds: string[];
-  loadBalancerSecurityGroupId: string;
-  loadBalancerTargetGroupArn: string;
-  secretsmanagerSecretArn: string;
-  fileStorageS3Endpoint: string;
-  fileStorageS3BucketName: string;
-  logsGroup: string;
-  vpcId: string;
+  subnets: vpc.Subnet[];
+  loadBalancerSecurityGroup: vpc.SecurityGroup;
+  loadBalancerTargetGroup: elb.AlbTargetGroup;
+  secret: secretsmanager.SecretsmanagerSecret;
+  logsGroup: cloudwatch.CloudwatchLogGroup;
+  vpc: vpc.Vpc;
   dockerApplications: DockerApplications;
   applicationBaseUrl: string;
 }
@@ -20,3 +33,11 @@ export interface EcsComponentPublicConfig {
 
 export type EcsComponentConfig = EcsComponentInternalConfig &
   EcsComponentPublicConfig;
+
+export interface EcsComponentArtifacts {
+  ecsCluster: ecs.EcsCluster;
+  executionRole: ExecutionRoleComponent;
+  privateDns: PrivateDnsComponent;
+  reverseProxy: ReverseProxyComponent;
+  reverseProxyTaskRole: ReverseProxyTaskRoleComponent;
+}