+++ libs/terraform/service/ecs/src/index.ts
@@ -1,3 +1,1 @@
-export * from './lib/containers';
-export * from './lib/interfaces';
-export * from './lib/services';
+export * from './lib';