+++ libs/terraform/service/ecs/src/lib/index.ts
@@ -1,0 +1,4 @@
+export * from './services';
+export * from './containers';
+export * from './configs';
+export * from './interfaces';