+++ libs/terraform/service/postgres/src/index.ts
@@ -1,3 +1,1 @@
-export * from './lib/interfaces';
-export * from './lib/postgres.component';
-export * from './lib/postgres.container';
+export * from './lib';