+++ libs/terraform/service/pwa/src/lib/index.ts
@@ -1,0 +1,4 @@
+export * from './interfaces';
+export * from './configs';
+export * from './pwa.component';
+export * from './pwa.container';